Remaking the packages and refactoring
